refactor: move masternode mode logic to `instantsend/signer.{cpp,h}`
Review with `git log -p -n1 --color-moved=dimmed_zebra`.
